integrating optional 1 (all solutions)
