+++ group/assignment-5/unknown_alphabet.py
@@ -1,51 +1,15 @@
 """Given a lexicon of all words in lexicographic order of a fictional Language
-find the alphabet of that Language
+find all alphabets of that Language
 """
-from typing import List, DefaultDict, Set
+from typing import List, Dict, DefaultDict, Set
 from collections import defaultdict
 
-def topological_sort(adjacency_list: DefaultDict[str, List[str]]) -> List[str]:
-    """Topological sort of a graph using the Kahn's algorithm.
-    Inputs:
-    a graph as an adjacency list
-    Returns:
-    a list containing the topological sort of our graph
-    """
-    
-    # Calculate the input degree for each vertex
-    in_degree = {}
-    for v, adjacents in [(vertex, adjacency_list[vertex]) for vertex in adjacency_list.keys()]:
-        in_degree.setdefault(v, 0)
-        for adj in adjacents:
-            in_degree[adj] = in_degree.get(adj, 0) + 1
-    
-    # Enqueue all vertices with input degree 0        
-    queue = {v for v, count in in_degree.items() if count == 0}
-    result = []
-    
-    # One by one dequeue vertices and enqueue adjacents if input degree becomes 0 
-    while queue:
-        v = queue.pop()
-        result.append(v)
-        
-        for adjacent in adjacency_list.get(v, []):
-            in_degree[adjacent] -= 1
-            
-            if in_degree[adjacent] == 0:
-                queue.add(adjacent)
-    
-    # Check for cycles in the graph
-    if len(result) != len(in_degree):
-        raise SyntaxError(f"The graph provided is not a DAG")
-    else:
-        return result
 
-def parse_lexicon(lexicon: List[List[str]]) -> DefaultDict[str, List[str]]:
+def parse_lexicon(lexicon: List[str]) -> DefaultDict[str, List[str]]:
     """Parse lexicon (collection of words in a fictional language)
     into an adjacency list of characters where an edge a -> b indicates that a goes before b
     """
     adj_list = defaultdict(list, {letter: [] for letter in set("".join(lexicon))})
-    #set([chr for word in lexicon for chr in word])
 
     for w1, w2 in zip(lexicon[:-1], lexicon[1:]):
         for l1, l2 in zip(w1, w2):
@@ -55,7 +19,43 @@ def parse_lexicon(lexicon: List[List[str]]) -> DefaultDict[str, List[str]]:
 
     return adj_list
 
-def get_alphabet(lexicon: List[List[str]]) -> List[str]:
-    """Obtain the alphabet in lexicographic order of a fictional language
+
+def topological_sort(adjacency_list: Dict[str, List[str]], in_degree: Dict, result: List[str]):
+    """Topological sort of a graph using the Kahn's algorithm.
+    Inputs:
+    a graph as an adjacency list
+    in_degree as a dictionary with input degree of every vertex
+    result as a prominent alphabet
+    Returns:
+    a generator which returns tuples containing a topological sort of the graph
     """
-    return topological_sort(parse_lexicon(lexicon))
+    if not in_degree:
+        # Calculate the input degree for each vertex
+        in_degree = {}
+        for v, adjacents in [(vertex, adjacency_list[vertex]) for vertex in adjacency_list.keys()]:
+            in_degree.setdefault(v, 0)
+            for adj in adjacents:
+                in_degree[adj] = in_degree.get(adj, 0) + 1
+
+    queue = {v for v, count in in_degree.items() if count == 0}
+
+    for v in queue:
+        in_degree[v] -= 1
+        result.append(v)
+        yield from topological_sort(adjacency_list,
+                                    {a:(in_degree[a]-1 if a in adjacency_list[v] else in_degree[a])
+                                    for a in in_degree.keys()},
+                                    result)
+        if all(in_degree[key] == -1 for key in in_degree.keys()):
+            yield tuple(result.copy())
+        
+        result.remove(v)
+        in_degree[v] += 1
+
+
+def get_alphabets(lexicon: List[str]) -> Set[List[str]]:
+    """
+    Returns a set of all alphabets for a given lexicon
+    """
+    alphabet_components = parse_lexicon(lexicon)
+    return {alphabet for alphabet in topological_sort(alphabet_components, in_degree=dict(), result=list())}
